Moved common utils to ckeditor5/dev/lib
