+++ dev/lib/log.js
