+++ dev/lib/tools.js
