+++ dev/tasks/dev/tasks/create-package.js
@@ -6,10 +6,10 @@
 'use strict';
 
 const inquiries = require( '../utils/inquiries' );
-const git = require( '../utils/git' );
-const tools = require( '../utils/tools' );
+const git = require( '../../../lib/git' );
+const tools = require( '../../../lib/tools' );
 const path = require( 'path' );
-const log = require( '../utils/log' );
+const log = require( '../../../lib/log' );
 
 /**
  * 1. Ask for new package name.
