+++ dev/tasks/test/tasks.js
@@ -10,7 +10,7 @@ const chai = require( 'chai' );
 const filterBy = require( 'gulp-filter-by' );
 const filter = require( 'gulp-filter' );
 const sinon = require( 'sinon' );
-const devTools = require( '../dev/utils/tools' );
+const devTools = require( '../../lib/tools' );
 const semver = require( 'semver' );
 const buildUtils = require( '../build/utils' );
 const benderConfig = require( '../../../bender' );
