+++ dev/tests/dev/log.js
@@ -8,7 +8,7 @@
 'use strict';
 
 const sinon = require( 'sinon' );
-const log = require( '../../tasks/dev/utils/log' );
+const log = require( '../../lib/log' );
 
 describe( 'log', () => {
 	it( 'should log using configured functions', () => {
